+++ internal/controller/meta/0module.go
@@ -1,0 +1,13 @@
+package meta
+
+import (
+	"go.uber.org/fx"
+)
+
+func Module() fx.Option {
+	return fx.Module("controllers.meta", fx.Invoke(
+		RegisterMeta,
+		RegisterIndex,
+		RegisterAdmin,
+	))
+}
