--- internal/controller/v3/0entry.go
@@ -1,17 +1,0 @@
-package v3
-
-import (
-	"go.uber.org/fx"
-)
-
-func Module() fx.Option {
-	opts := []fx.Option{
-		fx.Invoke(
-			RegisterItem,
-			RegisterLive,
-			RegisterStage,
-			RegisterZone,
-		),
-	}
-	return fx.Module("controllers.v3", opts...)
-}
