+++ internal/controller/v3/0module.go
@@ -1,0 +1,14 @@
+package v3
+
+import (
+	"go.uber.org/fx"
+)
+
+func Module() fx.Option {
+	return fx.Module("controllers.v3", fx.Invoke(
+		RegisterItem,
+		RegisterLive,
+		RegisterStage,
+		RegisterZone,
+	))
+}
