--- internal/repo/0entry.go
@@ -1,31 +1,0 @@
-package repo
-
-import (
-	"go.uber.org/fx"
-)
-
-func Module() fx.Option {
-	opts := []fx.Option{
-		fx.Provide(
-			NewItem,
-			NewZone,
-			NewAdmin,
-			NewStage,
-			NewNotice,
-			NewAccount,
-			NewActivity,
-			NewDropInfo,
-			NewProperty,
-			NewTimeRange,
-			NewDropReport,
-			NewRejectRule,
-			NewDropPattern,
-			NewTrendElement,
-			NewDropReportExtra,
-			NewDropMatrixElement,
-			NewDropPatternElement,
-			NewPatternMatrixElement,
-		),
-	}
-	return fx.Module("repo", opts...)
-}
