+++ internal/repo/0module.go
@@ -1,0 +1,28 @@
+package repo
+
+import (
+	"go.uber.org/fx"
+)
+
+func Module() fx.Option {
+	return fx.Module("repo", fx.Provide(
+		NewItem,
+		NewZone,
+		NewAdmin,
+		NewStage,
+		NewNotice,
+		NewAccount,
+		NewActivity,
+		NewDropInfo,
+		NewProperty,
+		NewTimeRange,
+		NewDropReport,
+		NewRejectRule,
+		NewDropPattern,
+		NewTrendElement,
+		NewDropReportExtra,
+		NewDropMatrixElement,
+		NewDropPatternElement,
+		NewPatternMatrixElement,
+	))
+}
