--- internal/service/0entry.go
@@ -1,36 +1,0 @@
-package service
-
-import (
-	"go.uber.org/fx"
-)
-
-func Module() fx.Option {
-	opts := []fx.Option{
-		fx.Provide(
-			NewItem,
-			NewZone,
-			NewStage,
-			NewGeoIP,
-			NewTrend,
-			NewAdmin,
-			NewHealth,
-			NewNotice,
-			NewReport,
-			NewAccount,
-			NewFormula,
-			NewActivity,
-			NewDropInfo,
-			NewShortURL,
-			NewTimeRange,
-			NewSiteStats,
-			NewDropMatrix,
-			NewDropReport,
-			NewTrendElement,
-			NewPatternMatrix,
-			NewDropMatrixElement,
-			NewDropPatternElement,
-			NewPatternMatrixElement,
-		),
-	}
-	return fx.Module("service", opts...)
-}
