+++ internal/service/0module.go
@@ -1,0 +1,33 @@
+package service
+
+import (
+	"go.uber.org/fx"
+)
+
+func Module() fx.Option {
+	return fx.Module("service", fx.Provide(
+		NewItem,
+		NewZone,
+		NewStage,
+		NewGeoIP,
+		NewTrend,
+		NewAdmin,
+		NewHealth,
+		NewNotice,
+		NewReport,
+		NewAccount,
+		NewFormula,
+		NewActivity,
+		NewDropInfo,
+		NewShortURL,
+		NewTimeRange,
+		NewSiteStats,
+		NewDropMatrix,
+		NewDropReport,
+		NewTrendElement,
+		NewPatternMatrix,
+		NewDropMatrixElement,
+		NewDropPatternElement,
+		NewPatternMatrixElement,
+	))
+}
